Added: Claim command
The main command for the whole event. This command can be run every 24h

Players will be able to claim a random 10-20 eggs for the their profiles
At times users might not be rewarded and sometimes they might lose eggs (5 less than what they would have earnt)

If a user claims each day they will earn up a streak. But if they miss a day the streak is reset
Every 7 days the user is rewarded with 50 eggs